[protocol] Update for garbage collection
After further analysis, garbage collection could be implemented much
more easily with a few protocol tweaks.

Firstly, remove rollover from the protocol by increasing the sequence
number space to 128 bits. If the sequence numbers are exhausted then a
new user or network identity can be allocated.

Secondly, the sender sequence number is not meaningful since the
sender cannot be relied upon to pick a good (monotonic + unique)
sequence number. Instead turn the protocol into a three way handshake
where the sequence number is assigned by the receiver, the sender
signs the transaction onto that sequence number, and then the receiver
wraps the entire package in an enclosing signature.

Lastly, allow the receiver to specify the previously valid sequence
number to allow for closing of holes in the record for garbage
collection. Holes could occur if a receiver assigns a sequence number
but the sender then fails. By allowing for holes the receiver can
continue to assign new sequence numbers to other exchanges while the
failed exchange is timing out.