[protocol][crdt-tp][ccn-server] Treat nets and users as entities
Now that the protocol has simplified and become more generic, it is
actually easier to treat all entities operating in the network the
same with respect to CRDT transactions. This allows using the same
behavior for both top-up and spend, just with different assignments of
who is the sender and who is the receiver.